Made a little modification for .NET 2.0 compatibility.
git-svn-id: https://cuyahoga.svn.sourceforge.net/svnroot/cuyahoga/trunk@452 cb474c29-17db-0310-ba2e-db2fec5ab780

+++ ServerControls/Pager.cs
@@ -480,12 +480,12 @@ namespace Cuyahoga.ServerControls
 
 		private int GetCurrentPageGroupIndex()
 		{
-			return (int)Math.Floor(this.CurrentPageIndex / this.MaxDisplayPages);
+			return (int)Math.Floor((double)(this.CurrentPageIndex / this.MaxDisplayPages));
 		}
 
 		private int GetTotalPageGroups()
 		{
-			return (int)Math.Ceiling(this.TotalPages / this.MaxDisplayPages);
+			return (int)Math.Ceiling((double)(this.TotalPages / this.MaxDisplayPages));
 		}
 
 		#endregion
